ecommerce/requirements.txt
black==21.11b1
coverage==6.2
Django==3.2.9
django-environ==0.8.1
django-extensions==3.1.2
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
ipython==7.30.1
isort==5.10.1
psycopg2-binary==2.9.2
unimport==0.9.2
Werkzeug==2.0.2
whitenoise==5.3.0
